update ER type to include transfers field.
Marked a TODO where necessary to fetch the data from Runtime when we have XDM enabled
