+++ crates/sp-domains-fraud-proof/src/bundle_equivocation.rs
@@ -191,6 +191,7 @@ mod test {
                     execution_trace: vec![],
                     execution_trace_root: Default::default(),
                     block_fees: Default::default(),
+                    transfers: Default::default(),
                 },
                 estimated_bundle_weight: Default::default(),
                 bundle_extrinsics_root: Default::default(),
